Remove dependency of site plugin on versioning plugin
