+++ kernel/src/main/scala/org/typelevel/sbt/TypelevelKernelPlugin.scala
@@ -16,10 +16,10 @@
 
 package org.typelevel.sbt
 
-import sbt._
+import sbt.*
 import sbt.plugins.JvmPlugin
-
-import Keys._
+import Keys.*
+import org.typelevel.sbt.kernel.{GitHelper, V}
 
 object TypelevelKernelPlugin extends AutoPlugin {
 
@@ -37,6 +37,36 @@ object TypelevelKernelPlugin extends AutoPlugin {
           BasicCommands.addAlias(BasicCommands.removeAlias(s, name), name, contents)
         }
       })
+
+    private[sbt] lazy val currentReleaseImpl = Def.setting {
+      // some tricky logic here ...
+      // if the latest release is a pre-release (e.g., M or RC)
+      // and there are no stable releases it is bincompatible with,
+      // then for all effective purposes it is the current release
+
+      val release = previousReleasesImpl.value match {
+        case head :: tail if head.isPrerelease =>
+          tail
+            .filterNot(_.isPrerelease)
+            .find(head.copy(prerelease = None).mustBeBinCompatWith(_))
+            .orElse(Some(head))
+        case releases => releases.headOption
+      }
+
+      release.map(_.toString)
+    }
+
+    // latest tagged release, including pre-releases
+    private[sbt] lazy val currentPreReleaseImpl = Def.setting {
+      previousReleasesImpl.value.headOption.map(_.toString)
+    }
+
+    private[sbt] lazy val previousReleasesImpl = Def.setting {
+      val currentVersion = V(version.value).map(_.copy(prerelease = None))
+      GitHelper.previousReleases(fromHead = true, strict = false).filter { v =>
+        currentVersion.forall(v.copy(prerelease = None) <= _)
+      }
+    }
   }
 
   import autoImport._
